chore[Op#50985]: Extract flash message helper from application helper
